requirements.txt
Django~=5.2
emoji~=2.14.0
attrs~=24.2.0
argparse~=1.4.0
distro~=1.9.0
wheel~=0.43.0
cryptography~=42.0.7
utils~=1.0.2
pillow~=10.3.0
tornado~=6.4.1
Jinja2~=3.1.4
filelock~=3.14.0
setuptools~=65.5.0
packaging~=24.0
pyparsing~=3.1.2
typing_extensions~=4.12.2
Cython~=3.0.11
future~=1.0.0
PyMySQL~=1.1.1
openpyxl~=3.1.5
torch~=2.6.0
torchvision~=0.21.0
opencv-python~=4.11.0.86
numpy~=1.26.4
safetensors~=0.4.5
scikit-image~=0.25.2
requests~=2.31.0
deepseek-api~=0.1.0